# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "description": "REST API for creating, assigning, updating, and summarizing team tasks.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List tasks",
                "operationId": "listTasks",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "assignee_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "due_date",
                                "priority",
                                "status"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated task list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCollection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "View a task",
                "operationId": "showTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Replace task fields",
                "operationId": "replaceTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/summary": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "View task summary counts",
                "operationId": "taskSummary",
                "responses": {
                    "200": {
                        "description": "Task summary counts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSummary"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserSummary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mara Santos"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "mara.santos@example.com"
                    }
                },
                "type": "object"
            },
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Build task creation endpoint"
                    },
                    "description": {
                        "type": "string",
                        "example": "Validate required fields and persist creator and optional assignee.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "high",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-20",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-13T08:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-13T08:00:00.000000Z"
                    },
                    "creator": {
                        "$ref": "#/components/schemas/UserSummary"
                    },
                    "assignee": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserSummary"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Task"
                    }
                },
                "type": "object"
            },
            "TaskCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "nullable": true
                            },
                            "last": {
                                "type": "string",
                                "nullable": true
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "path": {
                                "type": "string",
                                "example": "http://localhost:8000/api/tasks"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "to": {
                                "type": "integer",
                                "example": 5,
                                "nullable": true
                            },
                            "total": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TaskStoreRequest": {
                "required": [
                    "created_by_id",
                    "title"
                ],
                "properties": {
                    "created_by_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "assignee_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "example": "Build task API",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "example": "Create the task management endpoints.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "high",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-20",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TaskUpdateRequest": {
                "properties": {
                    "assignee_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "example": "Build task API",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "example": "Create the task management endpoints.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "completed",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "high",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-20",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "TaskSummary": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 5
                    },
                    "overdue": {
                        "type": "integer",
                        "example": 1
                    },
                    "by_status": {
                        "type": "object",
                        "example": {
                            "pending": 2,
                            "completed": 1
                        },
                        "additionalProperties": {
                            "type": "integer"
                        }
                    },
                    "by_priority": {
                        "type": "object",
                        "example": {
                            "medium": 2,
                            "high": 1
                        },
                        "additionalProperties": {
                            "type": "integer"
                        }
                    },
                    "by_assignee": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "assignee": {
                                    "type": "string",
                                    "example": "Leo Reyes"
                                },
                                "total": {
                                    "type": "integer",
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Task CRUD and summary endpoints"
        }
    ]
}